Fixed bug in comparing when calculate "Compos" (double=>float lost digits after 8 in mantissa)  SD
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@1375 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/s_formula.cpp
@@ -719,7 +719,7 @@ void TFormula::Stm_line( int N, double *Sml, char *ICsym, short *ICval )
         }
         if( jj==-1 )
             Error( icsp, "E33FPrun: This is not a symbol of IComp!");
-        Sml[jj] += (float)aSC[i];
+        Sml[jj] += aSC[i];
         if( aVal[i] == SHORT_EMPTY/*I_EMPTY*/ )
             aVal[i] = ICval[jj];
     }
